basic color styling and a grid layout for the posts on blog pages
